支持Android、UWP、iOS，实现alterTable，updateRecord等方法，修正部分bug

+++ easySQLite/easySQLite/easySQLite/SqlRecord.h
@@ -33,7 +33,7 @@ private:
 	void initColumnValue(int column_index, char* value, field_type type);
 
 public:
-	int columnCount();
+	size_t columnCount();
 	Value* getValue(int column_index);
 	Value* getValue(string fieldName);
 	Value* getKeyIdValue();
